♻️ refactor(agents-migrator): move prefix-ID→UUID remapping into AgentsMigrator
Replace the one-shot AgentIdMigrationSeeder with a standalone
remapAgentPrefixIds() function called from AgentsMigrator.execute()
after the import transaction commits.

The seeder was a conceptual misuse of the seeding system (which is for
inserting initial/builtin data) and had no real-database test coverage
for the remapping logic.

- Export remapAgentPrefixIds() with its own PRAGMA FK OFF + transaction
- Call it in AgentsMigrator.execute() after COMMIT, before DETACH
- Add 6 integration tests (setupTestDatabase) ported from the deleted
  seeder test: prefix IDs → UUID, FK cascade, builtin IDs, idempotency,
  and PRAGMA foreign_key_check
- Update AgentsMigrator unit test to mock db.transaction()
- Remove AgentIdMigrationSeeder and its test
