+++ src/main/data/db/seeding/index.ts
@@ -1,5 +1,4 @@
 import type { ISeeder } from '../types'
-import { AgentIdMigrationSeeder } from './seeders/agentIdMigrationSeeder'
 import { PreferenceSeeder } from './seeders/preferenceSeeder'
 import { PresetProviderSeeder } from './seeders/presetProviderSeeder'
 import { TranslateLanguageSeeder } from './seeders/translateLanguageSeeder'
@@ -9,9 +8,4 @@ import { TranslateLanguageSeeder } from './seeders/translateLanguageSeeder'
  * To add a new seeder: create an ISeeder class, add it to this array.
  * No changes to DbService needed.
  */
-export const seeders: ISeeder[] = [
-  new PreferenceSeeder(),
-  new TranslateLanguageSeeder(),
-  new PresetProviderSeeder(),
-  new AgentIdMigrationSeeder()
-]
+export const seeders: ISeeder[] = [new PreferenceSeeder(), new TranslateLanguageSeeder(), new PresetProviderSeeder()]
